+++ packages/thumbprint-react/CHANGELOG.md
@@ -7,6 +7,7 @@
 -   [Patch] Rewrite `FormNote` component in TypeScript.
 -   [Patch] Rewrite `Grid` and `GridColumn` components in TypeScript.
 -   [Patch] Rewrite `LoaderDots` component in TypeScript.
+-   [Patch] Rewrite `Tooltip` component in TypeScript.
 
 ## 9.5.0 - 2019-09-24